+++ src/games/gcd.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { startGame, roundsGames } from '../index.js';
+import startGame from '../index.js';
 import { getRandomInt, gedDel } from '../utils.js';
 
 const maxIntGcd = 100;
@@ -16,26 +16,17 @@ const getGcd = (num1, num2) => {
   return nod;
 };
 
-const getQuestionsAnswers = () => {
-  const questions = [];
-  const answers = [];
-  let counter = 0;
-  do {
-    counter += 1;
-    const randNum1 = getRandomInt(maxIntGcd);
-    const randNum2 = getRandomInt(maxIntGcd);
-    const curQuestion = `Question: ${randNum1} ${randNum2}`;
-    const curAnswer = getGcd(randNum1, randNum2);
-    questions.push(curQuestion);
-    answers.push(curAnswer.toString());
-  } while (counter < roundsGames);
-  return [questions, answers];
+const getRoundGcd = () => {
+  const randNum1 = getRandomInt(maxIntGcd);
+  const randNum2 = getRandomInt(maxIntGcd);
+  const curQuestion = `Question: ${randNum1} ${randNum2}`;
+  const curAnswer = getGcd(randNum1, randNum2);
+  return [curQuestion, curAnswer.toString()];
 };
 
 const gcdGame = () => {
   const gameCondition = 'Find the greatest common divisor of given numbers';
-  const [arrQuestions, arrAnswers] = getQuestionsAnswers();
-  startGame(gameCondition, arrQuestions, arrAnswers);
+  startGame(gameCondition, getRoundGcd);
 };
 
 export default gcdGame;
