+++ src/games/progression.js
@@ -1,4 +1,4 @@
-import { startGame, roundsGames } from '../index.js';
+import startGame from '../index.js';
 import { getRandomInt } from '../utils.js';
 
 const maxIntProgression = 20;
@@ -23,27 +23,18 @@ const getProgression = (randNum, increment) => {
   return [strProgression, missedNum];
 };
 
-const getQuestionsAnswers = () => {
-  const questions = [];
-  const answers = [];
-  let counter = 0;
-  do {
-    counter += 1;
-    const randNum = getRandomInt(maxIntProgression);
-    const increment = getRandomInt(maxIntProgression);
-    const [progression, missedNum] = getProgression(randNum, increment);
-    const curQuestion = `Question:${progression}`;
-    const curAnswer = missedNum;
-    questions.push(curQuestion);
-    answers.push(curAnswer.toString());
-  } while (counter < roundsGames);
-  return [questions, answers];
+const getRoundProgression = () => {
+  const randNum = getRandomInt(maxIntProgression);
+  const increment = getRandomInt(maxIntProgression);
+  const [progression, missedNum] = getProgression(randNum, increment);
+  const curQuestion = `Question:${progression}`;
+  const curAnswer = missedNum;
+  return [curQuestion, curAnswer.toString()];
 };
 
 const progressionGame = () => {
   const gameCondition = 'What number is missing in the progression?';
-  const [arrQuestions, arrAnswers] = getQuestionsAnswers();
-  startGame(gameCondition, arrQuestions, arrAnswers);
+  startGame(gameCondition, getRoundProgression);
 };
 
 export default progressionGame;
